requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autopep8==2.3.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
dj-rest-auth==6.0.0
Django==4.2
django-allauth==0.63.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docopt==0.6.2
drf-yasg==1.21.7
executing==2.0.1
fastjsonschema==2.20.0
frozenlist==1.4.1
future==1.0.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.7
inflection==0.5.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
oauthlib==3.2.2
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
stripe==10.1.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
